requirements.txt
typer[all]>=0.9.0
docker>=7.0.0
psycopg2-binary>=2.9.9
flask>=3.0.0
faker>=24.0.0
rich>=13.0.0
